move dummy/noop components to tests dir
