+++ tests/Integration/Config/ComponentProvider/Metrics/AggregationResolverExplicitBucketHistogram.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Config\SDK\ComponentProvider\Metrics;
+namespace OpenTelemetry\Tests\Integration\Config\ComponentProvider\Metrics;
 
 use OpenTelemetry\Config\SDK\Configuration\ComponentProvider;
 use OpenTelemetry\Config\SDK\Configuration\ComponentProviderRegistry;
